Make pruneFieldMap a static function
